delete utils.go
