--- x/ecocredit/util/utils.go
@@ -1,17 +1,0 @@
-package util
-
-import (
-	"strings"
-	"unicode"
-)
-
-func FastRemoveWhitespace(str string) string {
-	var b strings.Builder
-	b.Grow(len(str))
-	for _, ch := range str {
-		if !unicode.IsSpace(ch) {
-			b.WriteRune(ch)
-		}
-	}
-	return b.String()
-}
